+++ src/frontend/js/ashley.ts
@@ -1,10 +1,12 @@
-import * as AshleyEditor from './components/AshleyEditor';
+import { init } from './components/AshleyEditor';
 import { renderEmojis } from './utils/emojis';
 
 // expose some modules to the global window object
 declare var window: any;
 window.Ashley = {
-  Editor: AshleyEditor,
+  Editor: {
+    init,
+  },
 };
 
 document.addEventListener('DOMContentLoaded', event => {